Cleanup code now leaves the cache directory untouched if it is stored in gcs.

+++ src/main/java/pt/ulisboa/tecnico/graph/stream/GraphStreamHandler.java
@@ -1225,8 +1225,11 @@ public abstract class GraphStreamHandler<R> implements Runnable {
         }
 
         if ( ! this.keepingCacheDirectory) {
-            final java.nio.file.Path cachePath = Paths.get(this.cacheDirectory);
-            GraphUtils.deleteFileOrFolder(cachePath);
+			
+			if( ! this.cacheDirectory.startsWith("gs:")) {
+				final java.nio.file.Path cachePath = Paths.get(this.cacheDirectory);
+				GraphUtils.deleteFileOrFolder(cachePath);
+			}
         }
 
         if ( ! this.keepingTemporaryDirectory) {
